Copy records during _initial_vals create
_initial_vals was seeded with a reference to the objects records dict,
which meant that it was not actually storing the initial values. The
route53 API requires DELETEs to provide the entire existing objects
values, so this bug was preventing record changing via .save() to work.

_initial_vals.records is now a copy of the records object, and the
change_existing_rrset test has been updated to use the appropriate
attribute for testing (the `values` attribute was not actually being
honored by the route53 API)

Additionally the is_modified check was trying to enumerate _initial_vals
inappropriately, .items() call added